+++ mytimer/__main__.py
@@ -39,6 +39,7 @@ def main():
         const=1)
     parser.add_argument('--countup', help='countup timer', nargs="?", const=1)
     parser.add_argument('--alarm', help='alarm', nargs="?", const=1)
+    parser.add_argument('--faces-list', help='list of available faces', nargs="?", const=1)
     parser.add_argument('--version', help='version', nargs="?", const=1)
     args = parser.parse_args()
     run_timer(args)