+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.10] - 2023-10-15
+- `--face-list` flag
+	1. `print_faces_list()`
 ## [0.9] - 2023-10-04
 ### Added
 - 4 new faces